Fix Finder "can't handle objects ..." error on Mac
Occasionally, `Finder` will fail with the following error on MacOS:

```
  Finder got an error: Handler can’t handle objects of this class ...
```

This error can be avoided by using the built-in `open` command in
MacOS, which is also quite lightweight as compared to `osascript`.
`open -a Finder -R ...`, as used in the patch, essentially has the
same effect as the original `osascript` command minus the Finder
error. Thanks.